Create CommitForm component
